series expansion in log space
Repeated application of the affine and activation functions leads to
large errors. Graphical investigation shows that these errors occur at
|x| > 1, which suggests growing numerical instabilities.

Modifications have been started to test-ps-expansion.ipynb to use the z
= exp(ln z) identity, so that multiplication of large numbers (prod_k
z_k) are replaced with sums-of-logs.

This is still very much a work in progress, but I wanted to get this
out so others can contribute as they have time/bandwidth.